Separate session code from sql module
